local: refactor default os encoding out from local backend into shared encoder lib
