+++ backend/local/tests_test.go
@@ -3,6 +3,8 @@ package local
 import (
 	"runtime"
 	"testing"
+
+	"github.com/rclone/rclone/lib/encoder"
 )
 
 // Test Windows character replacements
@@ -21,7 +23,7 @@ func TestCleanWindows(t *testing.T) {
 		t.Skipf("windows only")
 	}
 	for _, test := range testsWindows {
-		got := cleanRootPath(test[0], true, defaultEnc)
+		got := cleanRootPath(test[0], true, encoder.OS)
 		expect := test[1]
 		if got != expect {
 			t.Fatalf("got %q, expected %q", got, expect)
