+++ lib/encoder/os_other.go
@@ -1,0 +1,6 @@
+//+build !windows,!darwin
+
+package encoder
+
+// OS is the encoding used by the local backend for non windows platforms
+const OS = Base
