+++ lib/encoder/os_windows.go
@@ -1,10 +1,8 @@
 //+build windows
 
-package local
+package encoder
 
-import "github.com/rclone/rclone/lib/encoder"
-
-// This is the encoding used by the local backend for windows platforms
+// OS is the encoding used by the local backend for windows platforms
 //
 // List of replaced characters:
 //   < (less than)     -> '＜' // FULLWIDTH LESS-THAN SIGN
@@ -24,10 +22,10 @@ import "github.com/rclone/rclone/lib/encoder"
 // Also encode invalid UTF-8 bytes as Go can't convert them to UTF-16.
 //
 // https://docs.microsoft.com/de-de/windows/desktop/FileIO/naming-a-file#naming-conventions
-const defaultEnc = (encoder.Base |
-	encoder.EncodeWin |
-	encoder.EncodeBackSlash |
-	encoder.EncodeCtl |
-	encoder.EncodeRightSpace |
-	encoder.EncodeRightPeriod |
-	encoder.EncodeInvalidUtf8)
+const OS = (Base |
+	EncodeWin |
+	EncodeBackSlash |
+	EncodeCtl |
+	EncodeRightSpace |
+	EncodeRightPeriod |
+	EncodeInvalidUtf8)
